Backend Runtime Ready
In the backend, it now creates game rooms, when enought players in
queueing (needs 2 players) and sends gameStart to those players who
joined the game, and sends gameEnd when it ends (by disconnection or
scores). While in game server sends gameUpdate with newly calculated
game user and ball positions aswell as current scores.

Remaining to do:
 - Implement frontend
 - Actual game logic